Rename: 그룹 페이지 폴더 구족 수정
