+++ src/pages/group/default/components/management/create-question-modal/create-question-form/index.ts
@@ -1,0 +1,1 @@
+export { CreateGroupQuestionForm } from './CreateQuestionForm'
