+++ src/pages/group/default/components/management/create-question-modal/index.ts
@@ -1,0 +1,1 @@
+export { CreateGroupQuestionModal } from './CreateQuestionModal'
