+++ src/pages/group/default/components/management/invite-member-modal/index.ts
@@ -1,0 +1,1 @@
+export { InviteMemberModal } from './InviteMemberModal'
